Rendering nice fonts in Java can be such a pain.

+++ src/main/java/be/guntherdw/minecraft/customfont/PrivateFieldsCustomFont.java
@@ -24,7 +24,7 @@ public class PrivateFieldsCustomFont<P, T> extends PrivateFields<P, T> {
     public static final PrivateFieldsCustomFont<FontRenderer, int[]>                colorCode = new PrivateFieldsCustomFont<FontRenderer, int[]>(FontRenderer.class, CustomFontObf.colorCode);
     public static final PrivateFieldsCustomFont<GuiNewChat, List<ChatLine>>    chatcomponents = new PrivateFieldsCustomFont<GuiNewChat, List<ChatLine>>(GuiNewChat.class, CustomFontObf.chatcomponents);
     public static final PrivateFieldsCustomFont<GuiChat, GuiTextField>             inputField = new PrivateFieldsCustomFont<GuiChat, GuiTextField>(GuiChat.class, CustomFontObf.inputField);
-    public static final PrivateFieldsCustomFont<Gui, Integer>                          zLevel = new PrivateFieldsCustomFont<Gui, Integer>(Gui.class, CustomFontObf.zLevel);
+    public static final PrivateFieldsCustomFont<Gui, Float>                            zLevel = new PrivateFieldsCustomFont<Gui, Float>(Gui.class, CustomFontObf.zLevel);
 
     // public static final PrivateFieldsCustomFont<EnumChatFormatting, Character>       formattingCode = new PrivateFieldsCustomFont<EnumChatFormatting, Character>(EnumChatFormatting.class, WatsonObf.EnumChatFormatting_formattingCode);
 }
